6.14重启项目
我阅读了整个项目，重新理解了项目结构。
1. 我把原本放在main里面的代码分离出来，放在了my_main里面。
2. 现在启动小车是处在State=-1的状态，这个状态下会在OLED上显示左轮的目标速度和实时速度，还有P、I、D三个参数。可以通过Vofa上位机绘制出实时速度和目标速度的曲线，还可以通过Vofa上位机通过拖动来改变P、I、D参数的值，观察曲线变化来调参。
3. Motor里封装了最基础的左右马达的启动、停止函数，还有调速函数。
4. pid里定义了pid的结构体，创建了两个速度环的实例，还有一个角度环的实例。定义了速度环和角度环的pid计算函数。定义了pid控制函数。
5. mpu6050和HMC5883L模块定义了从这两个芯片读出数据的函数。
6. HMC5883L_Calibration模块是用于HMC5883L的软铁硬铁校准的模块。可以校准磁力计的数据，是使用磁力计前必须的步骤。
7. vofa模块是用于和上位机的vofa软件通信用的模块，这个模块里面包含了发送和接收部分，很不完成，只用于调试左轮速度环，后继有需要可以完善。
8. gray_track模块里定义了两个函数，一个是track、一个是track_error。track_error没有输入，是根据当前传感器状态计算误差，然后返回误差的函数，会被track函数调用，而track函数又会被pid_control调用。track函数是速度环逻辑，可以理解为速度环的计算函数+计算结果的应用。它会调用track_error，计算当前偏差，同时会储存上一次的偏差，然后根据这些偏差进行pid计算，然后把计算出来的结果通过tar_set来应用。
9. 速度环会收敛于结构体内设定的目标速度，而tar_set函数可以设定这个目标速度。角度环是把一个基础速度±pid计算出来的差速，应用tar_set函数来控制角度的。
10. 而gray_track模块里面也是用tar_set函数来控制左右转弯，所以在这个版本里，角度环和循迹环**不能共存**。
11. 上面提到的tar_set函数实际上是pid_set_tar_speed函数。
12. 磁力计和陀螺仪的数据解析：
/*
1.  ax、ay、az，MPU6050 加速度计三轴原始值，在MPU6050模块里定义和修改。

2.  gx、gy、gz，MPU6050 陀螺仪三轴原始值，在MPU6050模块里定义和修改。

3.  gyro_zero_z，Z 轴陀螺仪零漂平均值，用来修正 gz。在MPU6050里的零漂校准函数被修改，这个函数会在初始被调用。

4. roll_gyro,  pitch_gyro,  yaw_gyro，只靠陀螺仪积分得到的姿态角。陀螺仪积分的特点：短时间很灵敏、很平滑。长时间会漂移，
因为一点点零漂会被不断积分放大。yaw_gyro 理论上可以表示偏航角，但会随时间慢慢跑偏。
在mpu6050模块里定义，在main里面的数据计算函数里计算。优点是短时间响应快，缺点是长时间零漂。

5. roll_acc, pitch_acc, yaw_acc，由加速度计粗略计算出来的角度。roll_acc 有物理意义。
pitch_acc 有物理意义。yaw_acc 这个写法数学上能算出一个角，
但作为真实偏航角并不可靠。在mpu6050模块里定义，在main里面的数据计算函数里计算。优点是不会零漂，缺点是运动激烈的时候会有扰动，不准。

6. hmc_x, hmc_y, hmc_z，HMC5883L 三轴磁力计原始值。在HMC5883L里被定义，在main里面的数据计算函数里计算。

7. hmc_x_cal, hmc_y_cal, hmc_z_cal，校准后的磁力计值。在HMC5883L里被定义，在main里面的数据计算函数里计算。

8. yaw_hmc，由 atan2f(hmc_y_cal, hmc_x_cal) 算出的磁力计航向角。在HMC5883L里被定义，在main里面的数据计算函数里计算。

9. HMC5883L_Cali_Res，储存校准结果的结构体。HMC5883L 校准结果，包含 offset 和 scale。
在HMC5883L_Calibration里被定义，在HMC5883L_Calibration里定义的校准函数里函数里计算。

10.  offset_x/y/z	三轴偏移量，用来修正硬铁干扰。scale_x/y/z	三轴缩放系数，用来修正软铁/轴向比例差异。

11. 若没有校准，全局 HMC5883L_Cali_Res 默认全是 0。这样 scale_x/scale_y 也是 0，hmc_x_cal/hmc_y_cal 会被算成 0，yaw_hmc 很可能失去实际意义。
除非你后面改代码给它赋过校准参数。
*/

+++ bsp/gray_track.c
@@ -6,38 +6,39 @@
 #include "Motor.h"
 
 float last_err = 0;
-uint8_t lose_cnt = 0;
+uint16_t lose_cnt = 0;
+
 float Kp = 14;
 float Kd = 2;
 float Kpp = 1.5;
 float Kdd = 0.5;
 
 // 通过灰度传感器计算error
+// TODO:这个地方要不要把循迹环也封装成一个结构体那样
 float track_error(void)
 {		
 	float sum = 0;
 	int cnt = 0;
 
-	if (O1 == 0) { sum -= 1.5; cnt++; }
-	if (O2 == 0) { sum -= 0.3; cnt++; }
-	if (O3 == 0) { cnt++; }
-	if (O4 == 0) { sum += 0.3; cnt++; }
-	if (O5 == 0) { sum += 1.5; cnt++; }
+	if(O1 == 0) { sum -= 1.5; cnt++; }
+	if(O2 == 0) { sum -= 0.3; cnt++; }
+	if(O3 == 0) { cnt++; }
+	if(O4 == 0) { sum += 0.3; cnt++; }
+	if(O5 == 0) { sum += 1.5; cnt++; }
 
-	if (cnt == 0)
+	// 丢线处理
+	if(cnt == 0)
 	{
+		// track_error被track调用，track在pid_control里被调用，而pid_control在【20ms】的定时器中断里被调用，
+		// 所以每一次lose_cnt代表的是20ms。
 		lose_cnt++;
-//		if(lose_cnt >= 260)
-//		{
-//			total_left = 0;
-//			total_right = 0;
-//			State = 1;
-//			lose_cnt = 0;
-//		}
 		if(lose_cnt >= 50)
 		{
+			// 先停车
 			pid_set_tar_speed(0, 0);
+			// 把速度环的目标角度设置为初始状态的反方向，希望以此让小车出弯后直行
 			angle.target = ANGStra - 180;
+			// 重置左右轮累计距离
 			total_left = 0;
 			total_right = 0;
 			State = 1;
@@ -56,14 +57,17 @@ float track_error(void)
 	}
 }
 
+
 void track(void)
 {
-	double err = track_error();
-	double derr = err - last_err;
+	float err = track_error();
+	float derr = err - last_err;
 	last_err = err;
 
+	// 灰度偏差P修正 + 灰度偏差D修正 + 非线性大偏差修正 + 陀螺仪角速度修正
 	float out = Kp * err + Kd * derr + Kpp * (err*fabs(err)) + Kdd * (float)((float)gz - gyro_zero_z)/16.4f;
 
+	// TODO:这个地方的base可以设定为全局变量方便修改
 	int base = 40;
 	int right = base - (int)out;
 	int left  = base + (int)out;
